requirements.txt
tensorflow>=2.13
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
opencv-python>=4.8
pillow>=10.0
streamlit>=1.32
